Changes to make it build correctly

+++ include/revolution/types.h
@@ -14,8 +14,8 @@
 #include <decomp.h>
 #include <macros.h>
 
-typedef unsigned long u64;
-typedef signed long s64;
+typedef unsigned long long u64;
+typedef signed long long s64;
 
 typedef unsigned int u32;
 typedef signed int s32;
